refactor(calculator): 일급 컬렉션
